+++ lib/ui/auth/widgets/social_icon_button.dart
@@ -1,0 +1,32 @@
+import 'package:flutter/material.dart';
+
+class SocialIconButton extends StatelessWidget {
+  final Widget icon;
+  final VoidCallback onPressed;
+  final Color? color;
+
+  const SocialIconButton({
+    super.key,
+    required this.icon,
+    required this.onPressed,
+    this.color,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return InkWell(
+      onTap: onPressed,
+      borderRadius: BorderRadius.circular(32),
+      child: Container(
+        height: 56,
+        width: 56,
+        alignment: Alignment.center,
+        decoration: BoxDecoration(
+          border: Border.all(color: Theme.of(context).dividerColor),
+          shape: BoxShape.circle,
+        ),
+        child: icon,
+      ),
+    );
+  }
+}